requirements.txt
# Core RAG — LangChain 1.x series
langchain==1.2.10
langchain-community==0.4.1
langchain-anthropic==1.3.4
langchain-chroma==1.1.0

# Embeddings
sentence-transformers==3.4.1
langchain-huggingface>=1.2.0

# Data fetching + Utilities (let pip resolve compatible versions)
# curl-cffi: libcurl-backed HTTP that passes Cloudflare TLS fingerprinting
curl-cffi>=0.14.0
requests>=2.32.5
python-dotenv>=1.0.1
tqdm>=4.67.1
